requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
# Snowflake & Airflow providers
apache-airflow-providers-snowflake
apache-airflow-providers-dbt-cloud
dbt-core
dbt-snowflake

# Python packages for ingestion
snowflake-connector-python[pandas]
pandas
pyyaml
typing-extensions